fixed issues related to refactoring
