+++ .importlinter.strict
@@ -32,7 +32,7 @@ ignore_imports =
     config.webapp -> platform.observability.sentry_sdk
     # core -> integrations/tools (T-4 debt — llm_cli)
     core.llm.factory -> integrations.llm_cli.registry
-    core.llm.factory -> integrations.llm_cli.runner
+    core.llm.client_builders -> integrations.llm_cli.runner
     core.llm.transports.sdk.agent_clients -> integrations.llm_cli.runner
     core.llm.transports.sdk.agent_clients -> integrations.llm_cli.text
     core.llm_invoke_errors -> integrations.llm_cli.errors
