+++ core/llm/parsers/__init__.py
@@ -1,0 +1,1 @@
+"""Parsers for free-text LLM output into structured fields."""
